+++ src/transports/smooth/isobmff/index.ts
@@ -14,27 +14,19 @@
  * limitations under the License.
  */
 
+import createAudioInitSegment from "./create_audio_init_segment";
+import createVideoInitSegment from "./create_video_init_segment";
+import parseTfrf, {
+  IISOBMFFBasicSegment,
+} from "./parse_tfrf";
+import parseTfxd from "./parse_tfxd";
+import patchSegment from "./patch_segment";
+
 export {
-  createBox,
-  createBoxWithChildren,
-} from "./create_box";
-export {
-  getBox,
-  getBoxContent,
-  getBoxOffsets,
-  getUuidContent,
-} from "./get_box";
-export {
-  getMDAT,
-  getMDIA,
-  getTRAF,
-} from "./read";
-export {
-  getMDHDTimescale,
-  getPlayReadyKIDFromPrivateData,
-  getTrackFragmentDecodeTime,
-  getDurationFromTrun,
-  getSegmentsFromSidx,
-  ISidxSegment,
-  patchPssh,
-} from "./utils";
+  createAudioInitSegment,
+  createVideoInitSegment,
+  IISOBMFFBasicSegment,
+  parseTfrf,
+  parseTfxd,
+  patchSegment,
+};
